+++ R/assertions.R
@@ -61,6 +61,7 @@ assert_task = function(task, task_type = NULL, feature_types = NULL, task_proper
 #' @param tasks (list of [Task]).
 #' @rdname mlr_assertions
 assert_tasks = function(tasks, task_type = NULL, feature_types = NULL, task_properties = NULL, .var.name = vname(tasks)) {
+  assert_list(tasks, types = "Task")
   invisible(lapply(tasks, assert_task, task_type = task_type, feature_types = feature_types, task_properties = task_properties, .var.name = .var.name))
 }
 
@@ -108,6 +109,7 @@ test_matching_task_type = function(task_type, object, class) {
 #' @param learners (list of [Learner]).
 #' @rdname mlr_assertions
 assert_learners = function(learners, task = NULL, task_type = NULL, properties = character(), unique_ids = FALSE, .var.name = vname(learners)) {
+  assert_list(learners, types = "Learner")
   if (unique_ids) {
     ids = map_chr(learners, "id")
     if (!test_character(ids, unique = TRUE)) {
@@ -123,7 +125,7 @@ assert_task_learner = function(task, learner, param_values = NULL, cols = NULL) 
   pars = learner$param_set$get_values(type = "only_token", check_required = FALSE)
   # remove pars that are covered by param_values
   pars = pars[names(pars) %nin% names(param_values)]
-  if (length(pars) > 0) {
+  if (length(pars) > 0L) {
     error_config("%s cannot be trained with TuneToken present in hyperparameter: %s", format_angle_brackets(learner), str_collapse(names(pars)))
   }
   # check on class(learner) does not work with GraphLearner and AutoTuner
@@ -135,7 +137,7 @@ assert_task_learner = function(task, learner, param_values = NULL, cols = NULL) 
   }
 
   tmp = setdiff(task$feature_types$type, learner$feature_types)
-  if (length(tmp) > 0) {
+  if (length(tmp) > 0L) {
     error_input("%s has the following unsupported feature types: %s", format_angle_brackets(task), str_collapse(tmp))
   }
 
@@ -250,7 +252,7 @@ assert_measure = function(measure, task = NULL, learner = NULL, prediction = NUL
 
     if (measure$check_prerequisites != "ignore") {
       miss = setdiff(measure$task_properties, task$properties)
-      if (length(miss) > 0) {
+      if (length(miss) > 0L) {
         warning_config("Measure '%s' is missing properties %s of task '%s'",
           measure$id, str_collapse(miss, quote = "'"), task$id)
       }
@@ -273,7 +275,7 @@ assert_measure = function(measure, task = NULL, learner = NULL, prediction = NUL
 
     if (measure$check_prerequisites != "ignore") {
       miss = setdiff(measure$predict_sets, learner$predict_sets)
-      if (length(miss) > 0) {
+      if (length(miss) > 0L) {
 
         warning_config("Measure '%s' needs predict sets %s, but learner '%s' only predicted on sets %s",
           measure$id, str_collapse(miss, quote = "'"), learner$id, str_collapse(learner$predict_sets, quote = "'"))
@@ -311,6 +313,7 @@ assert_scorable = function(measure, task, learner, prediction = NULL, .var.name 
 #' @param measures (list of [Measure]).
 #' @rdname mlr_assertions
 assert_measures = function(measures, task = NULL, learner = NULL, .var.name = vname(measures)) {
+  assert_list(measures, types = "Measure")
   lapply(measures, assert_measure, task = task, learner = learner, .var.name = .var.name)
   if (anyDuplicated(ids(measures))) {
     error_input("Measures need to have unique IDs")
@@ -341,6 +344,7 @@ assert_resampling = function(resampling, instantiated = NULL, .var.name = vname(
 #' @param resamplings (list of [Resampling]).
 #' @rdname mlr_assertions
 assert_resamplings = function(resamplings, instantiated = NULL, .var.name = vname(resamplings)) {
+  assert_list(resamplings, types = "Resampling")
   invisible(lapply(resamplings, assert_resampling, instantiated = instantiated, .var.name = .var.name))
 }
 
@@ -378,8 +382,8 @@ assert_set = function(x, empty = TRUE, .var.name = vname(x)) {
 assert_range = function(range, .var.name = vname(range)) {
   assert_numeric(range, len = 2L, any.missing = FALSE, .var.name = .var.name)
 
-  if (diff(range) <= 0) {
-    error_input("Invalid range specified. First value (%f) must be greater than second value (%f)", range[1L], range[2L])
+  if (diff(range) < 0) {
+    error_input("Invalid range specified. First value (%f) must be less than or equal to second value (%f)", range[1L], range[2L])
   }
 
   invisible(range)