+++ NEWS.md
@@ -1,5 +1,7 @@
 # mlr3 (development version)
 
+* feat: Add `$native_model` active binding to `Learner` to access the model from the upstream package.
+
 # mlr3 1.3.0
 
 * feat: `Learner$predict()` can now add additional data to `PredictionClassif` and `PredictionRegr` objects via the `extra` field.